GPU-2196: OpenAPI generation tests
