+++ Jenkinsfile
@@ -80,6 +80,7 @@ spec:
     parameters {
         booleanParam(name: 'checkGitCommit',            defaultValue: false, description: 'If set to true, the current git revision is compared with the git revision of the last successful build. If they are equal, the build is skipped and env.BUILD_SKIPPED is set to true')
         booleanParam(name: 'runTests',                  defaultValue: true,  description: "Whether to run the unit tests. tests will be skipped for MR builds if there are no relevant changes.")
+        booleanParam(name: 'useOpenApiTestClient',      defaultValue: false, description: "Whether to use an autogenerated OpenAPI Jersey 3 based REST client for the tests.")
         booleanParam(name: 'runBaseLibTests',           defaultValue: false,  description: "Whether to run tests from the base-lib module.")
         booleanParam(name: 'skiporacle',                defaultValue: false, description: "Whether to skip base-lib tests with Oracle")
 		booleanParam(name: 'skipmssql',                 defaultValue: false, description: "Whether to skip base-lib tests with MSSQL")
@@ -174,6 +175,9 @@ spec:
                         if (!params.runBaseLibTests) {
                             mvnArguments += "-Dsurefire.baselib.excludedGroups=com.gentics.contentnode.tests.category.BaseLibTest"
                         }
+                        if (params.useOpenApiTestClient) {
+							mvnArguments += "-Dtest.client.use.openapi=true"
+						}
                         def skipDBs = ""
                         if (params.skiporacle) {
                             skipDBs += " -Dskip.oracle12.2=true"
