+++ cms-restapi/src/main/java/com/gentics/contentnode/rest/resource/parameter/TemplateListParameterBean.java
@@ -14,8 +14,8 @@ public class TemplateListParameterBean {
 	/**
 	 * node id for this folder - for use with multichannelling
 	 */
-	@QueryParam("nodeId")
-	public Integer nodeId;
+	@QueryParam("folderNodeId")
+	public Integer folderNodeId;
 
 	/**
 	 * true to only return inherited templates in the given node, false to only get local/localized templates, null to get local and inherited templates
@@ -37,8 +37,8 @@ public class TemplateListParameterBean {
 	@DefaultValue("false")
 	public boolean reduce = false;
 
-	public TemplateListParameterBean setNodeId(Integer nodeId) {
-		this.nodeId = nodeId;
+	public TemplateListParameterBean setFolderNodeId(Integer nodeId) {
+		this.folderNodeId = nodeId;
 		return this;
 	}
