+++ pom.xml
@@ -104,6 +104,7 @@
 		<assertj.version>3.27.7</assertj.version>
 		<junit5.version>5.8.1</junit5.version>
 		<hsqldb.version>2.7.2</hsqldb.version>
+		<swagger.version>2.2.4</swagger.version>
 
 		<surefire.version>3.1.2</surefire.version>
 		<surefire.forkcount>16</surefire.forkcount>
